Relocate the App.js file tocomponets
