+++ src/index.js
@@ -3,7 +3,7 @@ import ReactDOM from 'react-dom';
 import './index.css';
 import { createStore } from 'redux';
 import { Provider } from 'react-redux';
-import App from './App';
+import App from './components/App';
 import reportWebVitals from './reportWebVitals';
 import rootReducer from './reducers ';
